fix(terminal): avoid events messing up topline of focused terminal
Problem: topline of a focused terminal window may not tail to terminal output if
events scroll the window.

Solution: set the topline in terminal_check_cursor.

Only set the topline if it changed, so the TextChangedT test in
functional/autocmd/termxx_spec.lua continues to pass...

Note the final terminal_check_cursor when leaving terminal mode still sets
topline like we're in terminal mode, but I think it's reasonable given that the
logic for bumping the cursor one column was retained, as it does so relative to
the terminal's actual cursor position (overwriting changes from events).